Modularised db connection code
