+++ server.js
@@ -1,6 +1,9 @@
 const express = require('express');
 const mongoose = require('mongoose');
-const bodyParser = require('body-parser');
+const dotenv = require('dotenv');
+const morgan = require('morgan');
+const connectDB = require('./config/db');
+
 const app = express();
 app.use(express.json());
 
@@ -10,27 +13,11 @@ if (process.env.NODE_ENV === 'development') {
   app.use(morgan('dev'));
 }
 
-const dotenv = require('dotenv');
+connectDB();
 
 dotenv.config();
-dotenv.config({ path: './.env' });
-
-const DB = process.env.DATABASE.replace(
-  '<PASSWORD>',
-  process.env.DATABASE_PASSWORD
-);
-
-mongoose
-  .connect(DB, {
-    useNewUrlParser: true,
-    useCreateIndex: true,
-    useFindAndModify: false,
-    useUnifiedTopology: true,
-  })
-  .then(() => console.log('DB connection successful still'));
 
 app.use('./posts', routes);
-app.use(bodyParser.json());
   
 app.use('/api/auth', require('./routes/authRoutes'));
